restore isRTL, remove Provider
