+++ src/commonStyles.ts
@@ -7,8 +7,8 @@ export const HOUR_GUIDE_WIDTH = 50
 export const OVERLAP_OFFSET = Platform.OS === 'web' ? 20 : 8
 export const OVERLAP_PADDING = Platform.OS === 'web' ? 3 : 0
 
-export const commonStyles = StyleSheet.create({
-  eventCell: {
+export const eventCellCss = StyleSheet.create({
+  style: {
     zIndex: 100,
     start: 3,
     end: 3,
