+++ src/utility-types.ts
@@ -1,0 +1,11 @@
+interface _DeepPartialArray<T> extends Array<DeepPartial<T>> {}
+/** @private */
+type _DeepPartialObject<T> = { [P in keyof T]?: DeepPartial<T[P]> }
+
+export type DeepPartial<T> = T extends Function
+  ? T
+  : T extends Array<infer U>
+  ? _DeepPartialArray<U>
+  : T extends object
+  ? _DeepPartialObject<T>
+  : T | undefined
