+++ stories/mobile.stories.tsx
@@ -3,7 +3,6 @@ import React from 'react'
 import { Alert, View } from 'react-native'
 
 import { Calendar } from '../src'
-import { ThemeProvider } from '../src/theme/ThemeContext'
 import { AppHeader, HEADER_HEIGHT } from './components/AppHeader'
 import { events } from './events'
 import { useEvents } from './hooks'
@@ -101,9 +100,13 @@ storiesOf('Mobile', module)
     }, [])
     return (
       <View style={styles.mobile}>
-        <ThemeProvider value={{ direction: 'rtl' }}>
-          <Calendar style={styles.calendar} locale="he" height={MOBILE_HEIGHT} events={events} />
-        </ThemeProvider>
+        <Calendar
+          isRTL
+          style={styles.calendar}
+          locale="he"
+          height={MOBILE_HEIGHT}
+          events={events}
+        />
       </View>
     )
   })
